pyproject.toml
[project]
name = "VOPy"
# version = "0.0.1"
dynamic = ["version"]
authors = [
  { name="Y. Cahit Yıldırım", email="cahit.yildirim@bilkent.edu.tr" },
]
description = "A framework for black-box vector optimization"
readme = "README.md"
license = {file="LICENSE"}
requires-python = ">=3.10"
dependencies = [
    "botorch>=0.10.0,<0.12.0",
    "cvxpy",
    "scikit-learn",
    "scipy",
    "matplotlib",
]

[project.urls]
Homepage = "https://github.com/Bilkent-CYBORG/VOPy"
Documentation = "https://vopy.readthedocs.io/en/latest/"
Issues = "https://github.com/Bilkent-CYBORG/VOPy/issues"

[project.optional-dependencies]
dev = ["black", "usort", "flake8", "flake8-print"]
test = ["pre-commit", "pytest", "pytest-cov", "bandit"]
examples = ["jupyter"]
docs = ["pydocstyle", "sphinx_rtd_theme", "sphinx-notfound-page", "nbsphinx"]

[build-system]
requires = ["setuptools>=64", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["vopy"]
include-package-data = true

[tool.setuptools.package-data]
"vopy" = ["datasets/data/*"]

[tool.setuptools_scm]
# Should be present even if its empty.
local_scheme = "node-and-date"
write_to = "./vopy/version.py"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::UserWarning",
]

[tool.black]
line-length = 100

# [tool.bandit.assert_used]
# skips = ['*/test_*.py']
